move ISliderCallback into its own header file
